requirements.txt
Django==2.2.17
django-clear-cache==0.3
django-cors-headers==3.6.0
djangorestframework==3.12.2
djongo==1.3.0
dnspython==2.1.0
gunicorn==20.0.4
pymongo==3.11.2
python-dateutil==2.8.1
rest-meets-djongo==0.0.13
sqlparse==0.2.4
whitenoise==5.2.0
